accounting-python/pyproject.toml
[project]
name = "accounting-python"
version = "0.1.0"
description = ""
dependencies = [
    "numpy>=1.21.5",
    "pandas>=1.2.4",
    "pdfminer.six>=v20231228",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "pytest>=8.3.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pyright]
extraPaths = [".venv/lib/python3.12/site-packages/"]